+++ src/arch/aarch64/vgic/vgic.h
@@ -58,8 +58,8 @@
 typedef void (*virq_ack_fn_t)(uint64_t vcpu_id, int irq, void *cookie);
 
 void vgic_init();
-bool handle_vgic_maintenance(uint64_t vcpu_id);
-bool handle_vgic_dist_fault(uint64_t vcpu_id, uint64_t fault_addr, uint64_t fsr, seL4_UserContext *regs);
-bool handle_vgic_redist_fault(uint64_t vcpu_id, uint64_t fault_addr, uint64_t fsr, seL4_UserContext *regs);
-bool vgic_register_irq(uint64_t vcpu_id, int virq_num, virq_ack_fn_t ack_fn, void *ack_data);
-bool vgic_inject_irq(uint64_t vcpu_id, int irq);
+bool handle_vgic_maintenance(size_t vcpu_id);
+bool handle_vgic_dist_fault(size_t vcpu_id, uint64_t fault_addr, uint64_t fsr, seL4_UserContext *regs);
+bool handle_vgic_redist_fault(size_t vcpu_id, uint64_t fault_addr, uint64_t fsr, seL4_UserContext *regs);
+bool vgic_register_irq(size_t vcpu_id, int virq_num, virq_ack_fn_t ack_fn, void *ack_data);
+bool vgic_inject_irq(size_t vcpu_id, int irq);
